# Conflicts:
#	Example/RainSDKDemo/RainSDKDemo/Core/Services/RainSDKService.swift
#	Example/RainSDKDemo/RainSDKDemo/Presentation/PortalWithdraw/PortalWithdrawDemoViewModel.swift
#	Example/RainSDKDemo/RainSDKDemo/Presentation/TransferDemo/TransferDemoViewModel.swift
#	Sources/RainSDK/Core/RainSDKError.swift
#	Sources/RainSDK/RainSDK.swift
#	Sources/RainSDK/RainSDKManager+Helpers.swift
#	Sources/RainSDK/RainSDKManager.swift
#	Tests/RainSDKTests/SendTokenTests.swift
#	Tests/RainSDKTests/WalletAgnosticInitializationTests.swift
#	Tests/RainSDKTests/WithdrawCollateralTests.swift

+++ Sources/RainSDK/Protocols/RainWalletProviderCapabilities.swift
@@ -33,7 +33,7 @@ internal protocol RainSolanaTransfersProvider: Sendable {
   func sendSolanaNative(
     chainId: Int,
     to toAddress: String,
-    amount: Double
+    amount: Decimal
   ) async throws -> String
 
   /// Signs and broadcasts an SPL token transfer (TransferChecked instruction, auto-creating
@@ -50,7 +50,7 @@ internal protocol RainSolanaTransfersProvider: Sendable {
     chainId: Int,
     mintAddress: String,
     to toAddress: String,
-    amount: Double,
+    amount: Decimal,
     decimals: Int
   ) async throws -> String
 }
